# docs/swagger.yaml
definitions:
  controller.ResponseArticleList:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/model.ArticleInfo'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  controller.ResponseCategoryList:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        description: 数据
        items:
          $ref: '#/definitions/model.Category'
        type: array
      message:
        description: 提示信息
        type: string
    type: object
  controller.ResponseGetSingleArticle:
    properties:
      code:
        description: 业务响应状态码
        type: integer
      data:
        $ref: '#/definitions/model.ArticleDetail'
        description: 返回数据
      message:
        description: 提示信息
        type: string
    type: object
  model.ArticleDetail:
    properties:
      categoryId:
        type: integer
      commentCount:
        type: integer
      content:
        description: 文章内容
        type: string
      createTime:
        description: 创建时间
        type: object
      id:
        type: integer
      summary:
        description: 文章摘要
        type: string
      title:
        type: string
      updateTime:
        type: object
      username:
        type: string
      viewCount:
        type: integer
    type: object
  model.ArticleInfo:
    properties:
      categoryId:
        type: integer
      commentCount:
        type: integer
      createTime:
        description: 创建时间
        type: object
      id:
        type: integer
      summary:
        description: 文章摘要
        type: string
      title:
        type: string
      updateTime:
        type: object
      username:
        type: string
      viewCount:
        type: integer
    type: object
  model.Category:
    properties:
      categoryId:
        type: integer
      categoryName:
        type: string
      categoryNo:
        type: integer
      color:
        type: string
    type: object
info:
  contact: {}
paths:
  /home/article:
    get:
      consumes:
      - application/json
      parameters:
      - description: 查询条件
        in: query
        name: condition
        type: string
      - description: 分类id
        in: query
        name: categoryId
        type: integer
      - description: 页面数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页码范围
        in: query
        name: pageSize
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseArticleList'
      tags:
      - 获取所有的文章列表
  /home/article/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: 文章标题
        in: body
        name: title
        required: true
        schema:
          type: string
      - description: 文章梗概
        in: body
        name: summary
        schema:
          type: string
      - description: 所属分类id
        in: body
        name: categoryId
        required: true
        schema:
          type: integer
      - description: 文章内容
        in: body
        name: content
        required: true
        schema:
          type: string
      - description: 文章作者名称
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 浏览数
        in: body
        name: viewCount
        required: true
        schema:
          type: integer
      - description: 评论数
        in: body
        name: commentCount
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      tags:
      - 文章保存
  /home/category:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseCategoryList'
      tags:
      - 获取所有的分类
  /home/getArticleById:
    get:
      parameters:
      - description: 文章id
        in: query
        name: articleId
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseGetSingleArticle'
      tags:
      - 根据id获取单个文章信息
  /user/save:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户id
        in: body
        name: userId
        required: true
        schema:
          type: string
      - description: 用户姓名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 邮箱
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: 邮箱
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
      tags:
      - 用户信息保存
  /user/username/select:
    get:
      parameters:
      - description: 用户姓名
        in: query
        name: condition
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      tags:
      - 查询数据库是否有已经注册
swagger: "2.0"